feat: #2109 PR文申請でPR画像をプレビュー表示
